Move styling files, move components to components folder
