+++ src/Components/App.js
@@ -1,10 +1,12 @@
 import React, { Component } from 'react';
-import './App.css';
 import { Route } from 'react-router-dom';
-import { fetchBooks } from './apiCalls';
-import Navbar from './Navbar';
+import { fetchBooks } from '../apiCalls';
+// import Navbar from './Navbar';
 import MainShelf from './MainShelf';
-import Footer from './Footer';
+import SaveShelf from './SaveShelf';
+// import Footer from './Footer';
+import '../styles/App.css';
+
 
 
 class App extends Component {
@@ -20,7 +22,7 @@ class App extends Component {
 
   componentDidMount() {
     fetchBooks()
-      .then(data => this.setState({ bookData: data, filteredBooks: data, savedBooks: [] }))
+      .then(data => this.setState({ bookData: data.results, filteredBooks: data.results, savedBooks: [] }))
       .catch(error => {
         console.log("Oh no!", error)
       });
@@ -33,13 +35,14 @@ class App extends Component {
       <div className="App">
         <header>
           <h1>GÜT BOOKS</h1>
-          <Navbar />
         </header>
         <main className="shelf-display">
           <Route exact path='/' render={() => <MainShelf bookData={this.state.filteredBooks} />} />
-          <Route exact path="/my-shelf" />
+          <Route exact path="/my-shelf" render={() => {
+            <SaveShelf savedBooks={this.state.savedBooks}
+            />
+          }} />
         </main>
-        < Footer />
       </div>
     )
   }
