+++ src/Components/Footer.js
@@ -1,0 +1,7 @@
+
+
+
+
+
+
+export default Footer;
