+++ src/Components/MainShelf.js
@@ -1,25 +1,25 @@
 import React from 'react';
 import Books from './Books';
 import BookFilter from './BookFilter';
-import './MainShelf.scss';
+import '../styles/MainShelf.scss';
 
-const MainShelf = ({ bookData }) => {
+function MainShelf({ bookData }) {
     const bookSpines = bookData.map(book => {
         return (
             <Books className="book-spine" key={book.id}>
                 <h2>${book.title}</h2>
             </Books>
-        )
-    })
+        );
+    });
 
     return (
         <section className="shelf-wrapper">
-            < BookFilter />
+            <BookFilter />
             <section className="book-shelf">
                 {bookSpines}
             </section>
         </section>
-    )
+    );
 }
