+++ src/styles/Books.css
@@ -1,20 +1,4 @@
-$brown: #76511e;
-$turquoise: #3394bb;
-$emerald: #1b7d54;
-$gold: #ecec71;
-
-
-.book-shelf {
-    width: 60%;
-    height: 120%;
-    margin: 0 auto;
-    border: 10px brown solid;
-    overflow: hidden;
-    background-color: black;
-}
-
-
-.book-spine {
+.book {
     height: 200px;
     width: 40px;
     float: left;
@@ -27,15 +11,15 @@ $gold: #ecec71;
 }
 
 
-.book-spine:hover {
+.book:hover {
     transform: scale(1.25);
 }
 
 
-.book-spine h2 {
+.book h2 {
     padding: 0;
     font-size: 1em;
     transform-origin: 0% 0%;
     transform: rotate(0.25turn) translateY(-35px);
     width: 170px;
-  }
+}
